move refsols to refsol_day5.rs; finalize day5 exercises
