+++ src/refsols/mod.rs
@@ -5,3 +5,5 @@ pub mod refsol_day2;
 pub mod refsol_day3;
 
 pub mod refsol_day4;
+
+pub mod refsol_day5;
